removed user_config file
