+++ app/lib/clients/vault.rb
@@ -6,7 +6,6 @@ module Clients
     extend Clients::Vault::Entity
     extend Clients::Vault::EntityAlias
     extend Clients::Vault::Oidc
-    extend Clients::Vault::UserConfig
 
     class_attribute :token
